@@ -58,7 +58,8 @@
         "nette/caching": "^3.0.0",
         "nikic/php-parser": "^4.19.1",
         "phpmd/phpmd": "^2.6.0",
-        "symfony/test-pack": "^1.1"
+        "symfony/test-pack": "^1.1",
+        "symfony/process": "^6.4"
     },
     "suggest": {
         "phplist/web-frontend": "5.0.x-dev",
