@@ -41,7 +41,7 @@ class SubscriberListRepositoryTest extends KernelTestCase
         parent::setUp();
         $this->loadSchema();
 
-        $this->subject = self::getContainer()->get(SubscriberListRepository::class);
+        $this->subscriberListRepository = self::getContainer()->get(SubscriberListRepository::class);
         $this->administratorRepository = self::getContainer()->get(AdministratorRepository::class);
         $this->subscriberRepository = self::getContainer()->get(SubscriberRepository::class);
         $this->subscriptionRepository = self::getContainer()->get(SubscriptionRepository::class);
@@ -59,7 +59,7 @@ class SubscriberListRepositoryTest extends KernelTestCase
         $this->loadFixtures([SubscriberListFixture::class]);
 
         $id = 1;
-        $creationDate = new DateTime('2016-06-22 15:01:17');
+        $creationDate = new DateTime();
         $modificationDate = new DateTime();
         $name = 'News';
         $description = 'News (and some fun stuff)';
@@ -68,7 +68,7 @@ class SubscriberListRepositoryTest extends KernelTestCase
         $category = 'news';
 
         /** @var SubscriberList $model */
-        $model = $this->subject->find($id);
+        $model = $this->subscriberListRepository->find($id);
 
         self::assertSame($id, $model->getId());
         self::assertSimilarDates($creationDate, $model->getCreationDate());
@@ -88,7 +88,7 @@ class SubscriberListRepositoryTest extends KernelTestCase
         $subscriberListId = 1;
         $ownerId = 1;
         /** @var SubscriberList $model */
-        $model = $this->subject->find($subscriberListId);
+        $model = $this->subscriberListRepository->find($subscriberListId);
         $owner = $model->getOwner();
 
         self::assertInstanceOf(Administrator::class, $owner);
@@ -119,9 +119,9 @@ class SubscriberListRepositoryTest extends KernelTestCase
     public function testSavePersistsAndFlushesModel()
     {
         $model = new SubscriberList();
-        $this->subject->save($model);
+        $this->subscriberListRepository->save($model);
 
-        self::assertSame($model, $this->subject->find($model->getId()));
+        self::assertSame($model, $this->subscriberListRepository->find($model->getId()));
     }
 
     public function testFindByOwnerFindsSubscriberListWithTheGivenOwner()
@@ -129,9 +129,9 @@ class SubscriberListRepositoryTest extends KernelTestCase
         $this->loadFixtures([SubscriberListFixture::class, AdministratorFixture::class]);
 
         $owner = $this->administratorRepository->find(1);
-        $ownedList = $this->subject->find(1);
+        $ownedList = $this->subscriberListRepository->find(1);
 
-        $result = $this->subject->findByOwner($owner);
+        $result = $this->subscriberListRepository->findByOwner($owner);
 
         self::assertContains($ownedList, $result);
     }
@@ -141,32 +141,32 @@ class SubscriberListRepositoryTest extends KernelTestCase
         $this->loadFixtures([SubscriberListFixture::class, AdministratorFixture::class]);
 
         $owner = $this->administratorRepository->find(1);
-        $foreignList = $this->subject->find(2);
+        $foreignList = $this->subscriberListRepository->find(2);
 
-        $result = $this->subject->findByOwner($owner);
+        $result = $this->subscriberListRepository->findByOwner($owner);
 
         self::assertNotContains($foreignList, $result);
     }
 
     public function testFindByOwnerIgnoresSubscriberListFromOtherOwner()
     {
-        $this->loadFixtures([SubscriberListFixture::class, AdministratorFixture::class]);
+        $this->loadFixtures([SubscriberListFixture::class]);
 
         $owner = $this->administratorRepository->find(1);
-        $unownedList = $this->subject->find(3);
+        $unownedList = $this->subscriberListRepository->find(3);
 
-        $result = $this->subject->findByOwner($owner);
+        $result = $this->subscriberListRepository->findByOwner($owner);
 
         self::assertNotContains($unownedList, $result);
     }
 
     public function testFindsAssociatedSubscriptions()
     {
-        $this->loadFixtures([SubscriberListFixture::class, SubscriberFixture::class, SubscriptionFixture::class]);
+        $this->loadFixtures([SubscriptionFixture::class]);
 
         $id = 2;
         /** @var SubscriberList $model */
-        $model = $this->subject->find($id);
+        $model = $this->subscriberListRepository->find($id);
         $subscriptions = $model->getSubscriptions();
 
         self::assertFalse($subscriptions->isEmpty());
@@ -183,7 +183,7 @@ class SubscriberListRepositoryTest extends KernelTestCase
 
         $id = 2;
         /** @var SubscriberList $model */
-        $model = $this->subject->find($id);
+        $model = $this->subscriberListRepository->find($id);
         $subscribers = $model->getSubscribers();
 
         $expectedSubscriber = $this->subscriberRepository->find(1);
@@ -200,12 +200,12 @@ class SubscriberListRepositoryTest extends KernelTestCase
 
         $id = 2;
         /** @var SubscriberList $model */
-        $model = $this->subject->find($id);
+        $model = $this->subscriberListRepository->find($id);
 
         $numberOfAssociatedSubscriptions = count($model->getSubscriptions());
         self::assertGreaterThan(0, $numberOfAssociatedSubscriptions);
 
-        $this->subject->remove($model);
+        $this->subscriberListRepository->remove($model);
 
         $newNumberOfSubscriptions = count($this->subscriptionRepository->findAll());
         $numberOfRemovedSubscriptions = $initialNumberOfSubscriptions - $newNumberOfSubscriptions;
@@ -217,13 +217,13 @@ class SubscriberListRepositoryTest extends KernelTestCase
         $this->loadFixtures([SubscriberListFixture::class]);
 
         /** @var SubscriberList[] $allModels */
-        $allModels = $this->subject->findAll();
+        $allModels = $this->subscriberListRepository->findAll();
         $numberOfModelsBeforeRemove = count($allModels);
         $firstModel = $allModels[0];
 
-        $this->subject->remove($firstModel);
+        $this->subscriberListRepository->remove($firstModel);
 
-        $numberOfModelsAfterRemove = count($this->subject->findAll());
+        $numberOfModelsAfterRemove = count($this->subscriberListRepository->findAll());
         self::assertSame(1, $numberOfModelsBeforeRemove - $numberOfModelsAfterRemove);
     }
 }
