@@ -10,10 +10,10 @@ use PhpList\Core\Core\Bootstrap;
 use PhpList\Core\Core\Environment;
 use PhpList\Core\Routing\ExtraLoader;
 use PHPUnit\Framework\TestCase;
-use Sensio\Bundle\FrameworkExtraBundle\Routing\AnnotatedRouteControllerLoader;
+use Symfony\Bundle\FrameworkBundle\Routing\AttributeRouteControllerLoader;
+use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Config\Loader\LoaderResolver;
-use Symfony\Component\HttpKernel\Config\FileLocator;
-use Symfony\Component\Routing\Loader\AnnotationDirectoryLoader;
+use Symfony\Component\Routing\Loader\AttributeDirectoryLoader;
 use Symfony\Component\Routing\Loader\YamlFileLoader;
 use Symfony\Component\Routing\RouteCollection;
 
@@ -34,16 +34,20 @@ class ExtraLoaderTest extends TestCase
 
         $this->kernel = $bootstrap->getApplicationKernel();
         $this->kernel->boot();
-        $container = $this->kernel->getContainer();
 
-        /** @var FileLocator $locator */
-        $locator = $container->get('file_locator');
-//        $attributeLoader = new AttributeRouteControllerLoader(new SimpleAnnotationReader());
+        $locator = new FileLocator([
+            $this->kernel->getProjectDir() . '/config',
+            $this->kernel->getProjectDir() . '/composer.json',
+            $this->kernel->getProjectDir() . '/src/EmptyStartPageBundle/Controller/'
+        ]);
+
+        $attributeLoader = new AttributeRouteControllerLoader();
+        $attributeDirectoryLoader = new AttributeDirectoryLoader($locator, $attributeLoader);
 
         $loaderResolver = new LoaderResolver(
             [
                 new YamlFileLoader($locator),
-//                new AttributeDirectoryLoader($locator, $attributeLoader),
+                $attributeDirectoryLoader,
             ]
         );
 
@@ -57,7 +61,7 @@ class ExtraLoaderTest extends TestCase
         Bootstrap::purgeInstance();
     }
 
-    public function testLoadReturnsRouteCollection()
+    public function testLoadReturnsRouteCollection(): void
     {
         self::assertInstanceOf(RouteCollection::class, $this->subject->load('', 'extra'));
     }
